LayoutTests/media/media-source/media-source-real-webm-duration-after-append.html

<!DOCTYPE html>
<html>
<head>
    <title>media-source-real-webm-duration-after-append</title>
    <script src="webm-generator.js"></script>
    <script src="../video-test.js"></script>
    <script>
    var source;
    var sourceBuffer;

    async function runTest() {
        findMediaElement();

        if (!MediaSource.isTypeSupported(WebM.VIDEO_TYPE)) {
            consoleWrite('SKIPPED: ' + WebM.VIDEO_TYPE + ' is not supported');
            return;
        }

        source = new MediaSource();
        video.src = URL.createObjectURL(source);
        await waitFor(source, 'sourceopen');

        // Set explicit duration to 0 (WebM init has no duration hint)
        run('source.duration = 0');

        sourceBuffer = source.addSourceBuffer(WebM.VIDEO_TYPE);

        sourceBuffer.appendBuffer(WebM.initSegment({ tracks: [{ id: 1, type: 'video' }] }));
        await waitFor(sourceBuffer, 'updateend');
        testExpected('source.duration', 0);

        // Append 10 one-second samples — duration should update.
        // WebM needs multiple samples to infer per-frame duration from
        // timecode gaps (unlike MP4 which has explicit trun durations).
        var samples = [];
        for (var i = 0; i < 10; i++)
            samples.push({ duration: 1000, isSync: i === 0 });
        sourceBuffer.appendBuffer(WebM.mediaSegment({
            tracks: [{ id: 1, type: 'video', baseDecodeTime: 0, samples }]
        }));
        await waitFor(sourceBuffer, 'updateend');

        // Duration should update based on end of buffered data
        testExpectedEqualWithTolerance('source.duration', 10, 0.1);
    }

    window.addEventListener('load', event => {
        runTest().then(endTest).catch(failTest);
    });
    </script>
</head>
<body>
    <video></video>
</body>
</html>
