LayoutTests/media/media-source/media-source-real-webm-duration-after-append-expected.txt


EVENT(sourceopen)
RUN(source.duration = 0)
EVENT(updateend)
EXPECTED (source.duration == '0') OK
EVENT(updateend)
EXPECTED (source.duration == '10') OK
END OF TEST
